chore: Correction of inconsistencies, rearrangement of rows, code cleaning.
